Disabled certain test because of missing feature
In czech you need different version of precise value for past and future

* en: before 1 minute, 30 seconds
* cs: před minutou, 30 sekundami

* en: za 1 minute, 30 seconds
* cs: za minutu, 30 sekund

The difference between "minutou" and "minutu" as well as "sekundami" and
 "sekund" cannot be captured in the code at the moment. Current result
of "před 1 minuta, 30 sekund" and "za 1 minuta, 30 sekund" is perfectly
understandable, albeit gramatically incorrect.

I also added more test to show correct behavior. It felt wrong to write
tests for the incorrect behavior. So there are none.